requirements.txt
apache-airflow==2.8.1
apache-airflow-providers-postgres==5.7.1
psycopg2-binary==2.9.9
pandas==2.1.4
faker==22.0.0
python-dotenv==1.0.0
pytest==7.4.4
